+++ svd-rs/src/derive_from.rs
@@ -112,19 +112,11 @@ where
     T: DeriveFrom,
 {
     fn derive_from(&self, other: &Self) -> Self {
-        match (self, other) {
-            (Self::Single(info), Self::Single(other_info)) => {
-                Self::Single(info.derive_from(other_info))
-            }
-            (Self::Single(info), Self::Array(other_info, other_dim)) => {
-                let mut dim = other_dim.clone();
-                dim.dim_name = None;
-                Self::Array(info.derive_from(other_info), dim)
-            }
-            (Self::Array(info, dim), Self::Single(other_info))
-            | (Self::Array(info, dim), Self::Array(other_info, _)) => {
-                Self::Array(info.derive_from(other_info), dim.clone())
-            }
+        use std::ops::Deref;
+        let info = self.deref().derive_from(other.deref());
+        match self {
+            Self::Single(_) => Self::Single(info),
+            Self::Array(_, dim) => Self::Array(info, dim.clone()),
         }
     }
 }